User can fetch data for investments

+++ lib/data/models/investment/investment_fixed_fund.dart
@@ -16,7 +16,7 @@ class InvestmentFixedFund {
   PercentageInterest percentageInterest;
   String currentValue;
   bool isMatured;
-  int withdrawableValue;
+  double withdrawableValue;
   DateTime maturityDate;
 
   factory InvestmentFixedFund.fromJson(Map<String, dynamic> json) => InvestmentFixedFund(
